+++ server/channelserver/sys_time.go
@@ -14,4 +14,5 @@ func TimeAdjusted() time.Time    { return gametime.Adjusted() }
 func TimeMidnight() time.Time    { return gametime.Midnight() }
 func TimeWeekStart() time.Time   { return gametime.WeekStart() }
 func TimeWeekNext() time.Time    { return gametime.WeekNext() }
+func TimeMonthStart() time.Time  { return gametime.MonthStart() }
 func TimeGameAbsolute() uint32   { return gametime.GameAbsolute() }